requirements.txt
streamlit
openai
pymongo
pdfplumber
scikit-learn
python-dotenv